# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a REST API for users and posts.",
    "version": "1.0.7",
    "title": "The API",
    "contact": {
      "email": "myemail@gmail.com"
    }
  },
  "host": "0.0.0.0:3000",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all users",
        "description": "This can only be done by the logged in user.",
        "operationId": "getAllUsers",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "403": {
            "description": "Invalid token or not token pair"
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Search users by params",
        "description": "This can only be done by the logged in user.",
        "operationId": "searchByParams",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "required": true,
            "type": "string",
            "name": "search",
            "description": "The user with this search param (id or email) will be founded"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "The data is not valid OR Search parameter is missing"
          },
          "401": {
            "description": "User not found"
          },
          "403": {
            "description": "Invalid token or not token pair"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "definitions": {
    "UserResponse": {
      "type": "object",
      "required": [
        "_id",
        "email",
        "password",
        "createAt",
        "updateAt"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "format": "uuid",
          "description": "The unique identifier for a user"
        },
        "name": {
          "type": "string",
          "default": null
        },
        "email": {
          "type": "string"
        },
        "age": {
          "type": "number",
          "default": 18
        },
        "password": {
          "type": "string"
        },
        "createAt": {
          "type": "string",
          "format": "date-time"
        },
        "updateAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
  }
  }